# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/librairie/upload-url": {
      "post": {
        "description": "Génère une clé d'objet sous `librairie/<documentId>/` et une URL PUT présignée. Le client envoie ensuite le fichier directement à R2, puis fournit la clé (et pour la couverture, le même documentId lors du second appel) via `fichierKey`/`coverKey` au POST /librairie.",
        "operationId": "LibrairieController_uploadUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "URL présignée générée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDocumentResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Demander une URL présignée pour un document ou sa couverture",
        "tags": [
          "Librairie"
        ]
      }
    },
    "/librairie": {
      "post": {
        "description": "Crée une nouvelle entrée de document à partir d'une clé de fichier (et éventuellement de couverture) déjà uploadée sur R2.",
        "operationId": "LibrairieController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Créer un document",
        "tags": [
          "Librairie"
        ]
      },
      "get": {
        "description": "Retourne les documents paginés avec filtres optionnels.",
        "operationId": "LibrairieController_findAll",
        "parameters": [
          {
            "name": "title",
            "required": true,
            "in": "query",
            "description": "Titre du document à rechercher",
            "schema": {
              "example": "Le Petit Prince",
              "type": "string"
            }
          },
          {
            "name": "categorie",
            "required": false,
            "in": "query",
            "description": "Filtrer par catégorie",
            "schema": {
              "example": "Budget",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Page de résultats à récupérer",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Nombre de résultats par page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée de documents",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/DocumentResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Liste des documents",
        "tags": [
          "Librairie"
        ]
      }
    },
    "/librairie/public": {
      "get": {
        "description": "Retourne les documents paginés sans données sensibles. Aucune authentification requise.",
        "operationId": "LibrairieController_findAllPublic",
        "parameters": [
          {
            "name": "title",
            "required": true,
            "in": "query",
            "description": "Titre du document à rechercher",
            "schema": {
              "example": "Le Petit Prince",
              "type": "string"
            }
          },
          {
            "name": "categorie",
            "required": false,
            "in": "query",
            "description": "Filtrer par catégorie",
            "schema": {
              "example": "Budget",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Page de résultats à récupérer",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Nombre de résultats par page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée de documents publics",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PublicDocumentResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Liste publique des documents",
        "tags": [
          "Librairie"
        ]
      }
    },
...
    "/librairie/public/{id}": {
      "get": {
        "description": "Retourne les informations publiques d'un document sans données sensibles. Aucune authentification requise.",
        "operationId": "LibrairieController_findOnePublic",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifiant du document",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDocumentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "summary": "Détail public d'un document",
        "tags": [
          "Librairie"
        ]
      }
    },
    "/librairie/{id}": {
      "get": {
        "operationId": "LibrairieController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifiant du document",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Détail d'un document",
        "tags": [
          "Librairie"
        ]
      },
      "patch": {
        "operationId": "LibrairieController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifiant du document",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Modifier un document",
        "tags": [
          "Librairie"
        ]
      },
      "delete": {
        "operationId": "LibrairieController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifiant du document",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document supprimé"
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Supprimer un document",
        "tags": [
          "Librairie"
        ]
      }
    },
    "/librairie/{id}/file": {
      "get": {
        "description": "Redirige (302) vers l'URL publique R2 du fichier associé au document.",
        "operationId": "LibrairieController_getFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifiant du document",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirection vers le fichier"
          },
          "404": {
            "description": "Document non trouvé"
          }
        },
        "summary": "Télécharger le fichier",
        "tags": [
          "Librairie"
        ]
      }
    },
...
  },
  "info": {
    "title": "Citoyen+ API",
    "description": "API officielle de la plateforme Citoyen+ — application citoyenne de la Côte d'Ivoire. Permet la gestion des actualités, signalements citoyens, bibliothèque de documents et quiz éducatifs.",
    "version": "1.0",
    "contact": {
      "name": "Équipe Citoyen+",
      "url": "https://mec-ci.org",
      "email": "contact@mec-ci.org"
    }
  },
...
  "servers": [
    {
      "url": "https://api.mec-ci.org",
      "description": "Production"
    },
    {
      "url": "http://localhost:8081",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Nombre total d'éléments",
            "example": 100
          },
          "page": {
            "type": "number",
            "description": "Page courante",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Nombre d'éléments par page",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "description": "Nombre total de pages",
            "example": 10
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UploadDocumentRequestDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "Nom du fichier tel qu'il sera envoyé au client (utilisé pour son extension). Pour kind=cover, le front redimensionne et convertit l'image en WebP avant l'upload (ex. couverture.webp).",
            "example": "rapport-annuel.pdf"
          },
          "contentType": {
            "type": "string",
            "description": "Type MIME du fichier, figé dans la signature de l’URL présignée",
            "example": "application/pdf"
          },
          "kind": {
            "type": "string",
            "description": "Type de fichier demandé : le document PDF ou sa couverture",
            "enum": [
              "fichier",
              "cover"
            ],
            "example": "fichier"
          },
          "documentId": {
            "type": "string",
            "description": "Identifiant de document à réutiliser, pour rattacher la couverture au même dossier que le fichier déjà demandé. Omis lors du premier appel (celui du fichier) : un nouvel identifiant est alors généré.",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "filename",
          "contentType",
          "kind"
        ]
      },
      "UploadDocumentResponseDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Clé objet R2 générée côté serveur"
          },
          "uploadUrl": {
            "type": "string",
            "description": "URL PUT présignée vers laquelle envoyer le fichier"
          },
          "expiresIn": {
            "type": "number",
            "description": "Durée de validité de l'URL présignée, en secondes"
          },
          "documentId": {
            "type": "string",
            "description": "Identifiant du dossier R2 (et du futur document) : à repasser en documentId lors de l'upload de la couverture, et retrouvable dans fichierKey/coverKey lors de POST /librairie.",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "key",
          "uploadUrl",
          "expiresIn",
          "documentId"
        ]
      },
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Titre du document",
            "example": "Le Petit Prince"
          },
          "description": {
            "type": "string",
            "description": "Description du document",
            "example": "Un conte philosophique et poétique"
          },
          "categorie": {
            "type": "string",
            "description": "Catégorie du document",
            "example": "Budget"
          },
          "userId": {
            "type": "string",
            "description": "ID de l'utilisateur propriétaire du document",
            "example": 1
          },
          "fichierKey": {
            "type": "string",
            "description": "Clé R2 du fichier PDF, obtenue via POST /librairie/upload-url (kind: 'fichier')",
            "example": "librairie/550e8400-e29b-41d4-a716-446655440000/fichier.pdf"
          },
          "coverKey": {
            "type": "string",
            "description": "Clé R2 de la couverture, obtenue via POST /librairie/upload-url (kind: 'cover')",
            "example": "librairie/550e8400-e29b-41d4-a716-446655440000/cover.jpg"
          }
        },
        "required": [
          "title",
          "description",
          "userId",
          "fichierKey"
        ]
      },
      "DocumentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID du document",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "description": "Titre du document",
            "example": "Le Petit Prince"
          },
          "description": {
            "type": "object",
            "description": "Description du document",
            "example": "Un conte philosophique et poétique",
            "nullable": true
          },
          "categorie": {
            "type": "object",
            "description": "Catégorie du document",
            "example": "Budget",
            "nullable": true
          },
          "fileType": {
            "type": "string",
            "description": "Type de fichier",
            "example": ".pdf"
          },
          "fileUrl": {
            "type": "string",
            "description": "URL pour télécharger le fichier",
            "example": "/api/v1/librairie/550e8400-e29b-41d4-a716-446655440000/file"
          },
          "coverImage": {
            "type": "object",
            "description": "URL publique de la couverture",
            "example": "https://cdn.mec-ci.org/librairie/550e8400-e29b-41d4-a716-446655440000/cover.jpg",
            "nullable": true
          },
          "pageCount": {
            "type": "number",
            "description": "Nombre de pages du document (détecté automatiquement pour les PDF)",
            "nullable": true,
            "example": 42
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date d'upload",
            "example": "2026-01-24T10:30:00Z"
          },
          "uploadedBy": {
            "type": "object",
            "description": "Informations de l'utilisateur ayant uploadé"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "categorie",
          "fileType",
          "fileUrl",
          "coverImage",
          "pageCount",
          "uploadedAt",
          "uploadedBy"
        ]
      },
      "PublicDocumentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID du document",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "description": "Titre du document",
            "example": "Le Petit Prince"
          },
          "description": {
            "type": "object",
            "description": "Description du document",
            "nullable": true
          },
          "categorie": {
            "type": "object",
            "description": "Catégorie du document",
            "example": "Budget",
            "nullable": true
          },
          "fileType": {
            "type": "string",
            "description": "Type de fichier",
            "example": ".pdf"
          },
          "fileUrl": {
            "type": "string",
            "description": "URL pour télécharger le fichier"
          },
          "coverImage": {
            "type": "object",
            "description": "URL de la couverture",
            "nullable": true
          },
          "pageCount": {
            "type": "number",
            "description": "Nombre de pages du document (détecté automatiquement pour les PDF)",
            "nullable": true,
            "example": 42
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date d'upload"
          },
          "auteur": {
            "type": "string",
            "description": "Nom de l'auteur du document",
            "example": "Jean Dupont"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "categorie",
          "fileType",
          "fileUrl",
          "coverImage",
          "pageCount",
          "uploadedAt",
          "auteur"
        ]
      },
      "UpdateDocumentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Titre du document",
            "example": "Le Petit Prince"
          },
          "description": {
            "type": "string",
            "description": "Description du document",
            "example": "Un conte philosophique et poétique"
          },
          "categorie": {
            "type": "string",
            "description": "Catégorie du document",
            "example": "Budget"
          },
          "userId": {
            "type": "string",
            "description": "ID de l'utilisateur propriétaire du document",
            "example": 1
          },
          "fichierKey": {
            "type": "string",
            "description": "Clé R2 du fichier PDF, obtenue via POST /librairie/upload-url (kind: 'fichier')",
            "example": "librairie/550e8400-e29b-41d4-a716-446655440000/fichier.pdf"
          },
          "coverKey": {
            "type": "string",
            "description": "Clé R2 de la couverture, obtenue via POST /librairie/upload-url (kind: 'cover')",
            "example": "librairie/550e8400-e29b-41d4-a716-446655440000/cover.jpg"
          }
        }
      },
...
    }
  }
}